♻️ move `RecursivePartial` from `@datadog/js-core/assembly` to `@datadog/js-core/util`
`RecursivePartial` is a generic utility type with no dependency on the
hook assembly pattern, so `util` is a more appropriate home. The
`assembly` sub-path re-exports it for backward compatibility.
