+++ packages/browser-logs/src/domain/hooks.ts
@@ -1,7 +1,8 @@
 import type { RelativeTime } from '@datadog/js-core/time'
 import type { TelemetryEvent } from '@datadog/browser-core'
 import { createHook } from '@datadog/js-core/assembly'
-import type { Hook, RecursivePartial } from '@datadog/js-core/assembly'
+import type { Hook } from '@datadog/js-core/assembly'
+import type { RecursivePartial } from '@datadog/js-core/util'
 import type { LogsEvent } from '../logsEvent.types'
 
 export type DefaultLogsEventAttributes = RecursivePartial<LogsEvent>
