+++ packages/browser-rum-core/src/domain/hooks.ts
@@ -1,7 +1,8 @@
 import type { Duration, RelativeTime } from '@datadog/js-core/time'
 import type { TelemetryEvent } from '@datadog/browser-core'
 import { createHook } from '@datadog/js-core/assembly'
-import type { Hook, RecursivePartial } from '@datadog/js-core/assembly'
+import type { Hook } from '@datadog/js-core/assembly'
+import type { RecursivePartial } from '@datadog/js-core/util'
 import type { RumEvent } from '../rumEvent.types'
 import type { RawRumEvent } from '../rawRumEvent.types'
 import type { RumEventDomainContext } from '../domainContext.types'
